fix: Size IC3 indices and counts by what they index and count
Proof goals already stored their frame index as size_t but were handed
one as unsigned int, and the SyGuS per-cex records counted terms and
assumptions in unsigned. The random seed and generalization iteration
cap go the other way: smt-switch's reducer takes them as unsigned int,
so the options now hold that type and the parser rejects a value that
does not fit rather than wrapping it.

+++ engines/ic3base.h
@@ -148,7 +148,7 @@ class ProofGoalQueue
 
   void clear();
   void new_proof_goal(const IC3Formula & c,
-                      unsigned int t,
+                      size_t t,
                       const ProofGoal * n = NULL);
   ProofGoal * top();
   void pop();
